requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
celery==5.3.1
redis==5.0.0
gevent==23.9.1
httpx==0.25.2
flower==2.0.1
pydantic==2.4.2
python-dotenv==1.0.0